+++ libs/js_engine/Cargo.toml
@@ -11,7 +11,7 @@ dashmap = "5.3.4"
 deno_core = "=0.136.0"
 deno_ast = { version = "=0.15.0", features = ["transpiling"] }
 dprint-swc-ext = "=0.1.1"
-embassy_container_init = { path = "../embassy-container-init" }
+embassy_container_init = { path = "../embassy_container_init" }
 reqwest = { version = "0.11.11" }
 swc_atoms = "=0.2.11"
 swc_common = "=0.18.7"
