+++ libs/snapshot_creator/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "snapshot-creator"
+name = "snapshot_creator"
 version = "0.1.0"
 edition = "2021"
